src/server/mcp_hub/file_management/requirements.txt
fastmcp
python-dotenvtextract
python-docx2txt
ebooklib
beautifulsoup4
msg-extractor
python-pptx
xlrd
